Fixed Day/Night Cycle
Improved the Day and Night cycle with alternating the water lighting to be darker towards the night and brighter as the sun comes up.

Have also added in the option to have it represent your system time so you can play during the day in the real world and be the day and time of your system time.  Same for the night. This will be an optional toggle for the player.